docker/clb/mmapi/requirements.txt
Flask==1.1.1
pymongo==3.8.0
Flask_RESTful==0.3.7
Flask_Cors==3.0.8
ldap3==2.6.1
PyJWT==1.7.1
pyModbusTCP==0.1.8
pyopenssl==19.0.0
gunicorn==19.9.0